response 클래스 생성 dos를 통한 response를 담당하는 response 클래스 생성 response() 메소드가 responseHeader, responseBody 수행
